Improve code example in docs/tans.rst (#184)
* Add product_id to example in docs/tans.rst

* Fix initial tan request in example in docs/tans.rst

Before this commit, the example suggested that we should only ask for
a TAN after calling client.get_transactions(). This did not work (at least
for my bank Sparkasse Hannover). Now with this commit, the example is
proposing to ask for a TAN already in the beginning, as also suggested
in docs/trouble.rst.

Another side effect of this commit is that ask_for_tan() now also checks
for response.decoupled and presents a more reasonable string for input()
in the case of PushTAN.